+++ packages/faceted-search/CHANGELOG.md
@@ -28,7 +28,9 @@ Types of changes
 
 ### Fixed
 
+- [fixed](https://github.com/Talend/ui/pull/2665): sort alphabetically facets
 - [fixed](https://github.com/Talend/ui/pull/2662): values open up at start
+- [fixed](https://github.com/Talend/ui/pull/2666): Put a scrollbar for enum badge with a lot of values
 
 ## [0.3.0]